+++ routes/login.js
@@ -1,16 +1,9 @@
 'use strict';
+var SessionController = require('../Controller/SessionController');
+
 const express = require('express');
 const router = express.Router();
 
-class SessionController {
-  login(req, res) {
-    const from = req.query.from;
-    if (from) {
-      res.cookie('loginFrom', from, { expires: new Date(Date.now() + 600000)});
-    }
-    res.render('login');
-  }
-}
 
 let sessionController = new SessionController();
